+++ bindings/na/zoned/zoned.cpp
@@ -51,28 +51,45 @@ PYBIND11_MODULE(MQT_QMAP_MODULE_NAME, m, py::mod_gil_not_used()) {
                      return self.exportNAVizMachine();
                    });
 
+  //===--------------------------------------------------------------------===//
+  // Routing-agnostic Compiler
+  //===--------------------------------------------------------------------===//
   py::class_<na::zoned::RoutingAgnosticCompiler> routingAgnosticCompiler(
       m, "RoutingAgnosticCompiler");
-  routingAgnosticCompiler.def(
-      py::init([](const na::zoned::Architecture& arch,
-                  const std::string& logLevel, const bool useWindow,
-                  const size_t windowSize, const bool dynamicPlacement,
-                  const size_t parkingOffset, const bool warnUnsupportedGates)
-                   -> na::zoned::RoutingAgnosticCompiler {
-        na::zoned::RoutingAgnosticCompiler::Config config;
-        config.logLevel = spdlog::level::from_str(logLevel);
-        config.layoutSynthesizerConfig.placerConfig = {.useWindow = useWindow,
-                                                       .windowSize = windowSize,
-                                                       .dynamicPlacement =
-                                                           dynamicPlacement};
-        config.codeGeneratorConfig = {.parkingOffset = parkingOffset,
-                                      .warnUnsupportedGates =
-                                          warnUnsupportedGates};
-        return {arch, config};
-      }),
-      py::keep_alive<1, 2>(), "arch"_a, "log_level"_a = "warning",
-      "use_window"_a = true, "window_size"_a = 10, "dynamic_placement"_a = true,
-      "parking_offset"_a = 1, "warn_unsupported_gates"_a = true);
+  {
+    const na::zoned::RoutingAgnosticCompiler::Config defaultConfig;
+    routingAgnosticCompiler.def(
+        py::init([](const na::zoned::Architecture& arch,
+                    const std::string& logLevel, const double maxFillingFactor,
+                    const bool useWindow, const size_t windowSize,
+                    const bool dynamicPlacement, const size_t parkingOffset,
+                    const bool warnUnsupportedGates)
+                     -> na::zoned::RoutingAgnosticCompiler {
+          na::zoned::RoutingAgnosticCompiler::Config config;
+          config.logLevel = spdlog::level::from_str(logLevel);
+          config.schedulerConfig.maxFillingFactor = maxFillingFactor;
+          config.layoutSynthesizerConfig.placerConfig = {
+              .useWindow = useWindow,
+              .windowSize = windowSize,
+              .dynamicPlacement = dynamicPlacement};
+          config.codeGeneratorConfig = {.parkingOffset = parkingOffset,
+                                        .warnUnsupportedGates =
+                                            warnUnsupportedGates};
+          return {arch, config};
+        }),
+        py::keep_alive<1, 2>(), "arch"_a,
+        "log_level"_a = spdlog::level::to_short_c_str(defaultConfig.logLevel),
+        "max_filling_factor"_a = defaultConfig.schedulerConfig.maxFillingFactor,
+        "use_window"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.useWindow,
+        "window_size"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.windowSize,
+        "dynamic_placement"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.dynamicPlacement,
+        "parking_offset"_a = defaultConfig.codeGeneratorConfig.parkingOffset,
+        "warn_unsupported_gates"_a =
+            defaultConfig.codeGeneratorConfig.warnUnsupportedGates);
+  }
   routingAgnosticCompiler.def_static(
       "from_json_string",
       [](const na::zoned::Architecture& arch,
@@ -96,40 +113,66 @@ PYBIND11_MODULE(MQT_QMAP_MODULE_NAME, m, py::mod_gil_not_used()) {
         return self.getStatistics();
       });
 
+  //===--------------------------------------------------------------------===//
+  // Routing-aware Compiler
+  //===--------------------------------------------------------------------===//
   py::class_<na::zoned::RoutingAwareCompiler> routingAwareCompiler(
       m, "RoutingAwareCompiler");
-  routingAwareCompiler.def(
-      py::init([](const na::zoned::Architecture& arch,
-                  const std::string& logLevel, const bool useWindow,
-                  const size_t windowMinWidth, const double windowRatio,
-                  const double windowShare, const float deepeningFactor,
-                  const float deepeningValue, const float lookaheadFactor,
-                  const float reuseLevel, const size_t maxNodes,
-                  const size_t parkingOffset, const bool warnUnsupportedGates)
-                   -> na::zoned::RoutingAwareCompiler {
-        na::zoned::RoutingAwareCompiler::Config config;
-        config.logLevel = spdlog::level::from_str(logLevel);
-        config.layoutSynthesizerConfig.placerConfig = {
-            .useWindow = useWindow,
-            .windowMinWidth = windowMinWidth,
-            .windowRatio = windowRatio,
-            .windowShare = windowShare,
-            .deepeningFactor = deepeningFactor,
-            .deepeningValue = deepeningValue,
-            .lookaheadFactor = lookaheadFactor,
-            .reuseLevel = reuseLevel,
-            .maxNodes = maxNodes};
-        config.codeGeneratorConfig = {.parkingOffset = parkingOffset,
-                                      .warnUnsupportedGates =
-                                          warnUnsupportedGates};
-        return {arch, config};
-      }),
-      py::keep_alive<1, 2>(), "arch"_a, "log_level"_a = "warning",
-      "use_window"_a = true, "window_min_width"_a = 8, "window_ratio"_a = 1.0,
-      "window_share"_a = 0.6, "deepening_factor"_a = 0.8,
-      "deepening_value"_a = 0.2, "lookahead_factor"_a = 0.2,
-      "reuse_level"_a = 5.0, "max_nodes"_a = 50000000, "parking_offset"_a = 1,
-      "warn_unsupported_gates"_a = true);
+  {
+    const na::zoned::RoutingAwareCompiler::Config defaultConfig;
+    routingAwareCompiler.def(
+        py::init([](const na::zoned::Architecture& arch,
+                    const std::string& logLevel, const double maxFillingFactor,
+                    const bool useWindow, const size_t windowMinWidth,
+                    const double windowRatio, const double windowShare,
+                    const float deepeningFactor, const float deepeningValue,
+                    const float lookaheadFactor, const float reuseLevel,
+                    const size_t maxNodes, const size_t parkingOffset,
+                    const bool warnUnsupportedGates)
+                     -> na::zoned::RoutingAwareCompiler {
+          na::zoned::RoutingAwareCompiler::Config config;
+          config.logLevel = spdlog::level::from_str(logLevel);
+          config.schedulerConfig.maxFillingFactor = maxFillingFactor;
+          config.layoutSynthesizerConfig.placerConfig = {
+              .useWindow = useWindow,
+              .windowMinWidth = windowMinWidth,
+              .windowRatio = windowRatio,
+              .windowShare = windowShare,
+              .deepeningFactor = deepeningFactor,
+              .deepeningValue = deepeningValue,
+              .lookaheadFactor = lookaheadFactor,
+              .reuseLevel = reuseLevel,
+              .maxNodes = maxNodes};
+          config.codeGeneratorConfig = {.parkingOffset = parkingOffset,
+                                        .warnUnsupportedGates =
+                                            warnUnsupportedGates};
+          return {arch, config};
+        }),
+        py::keep_alive<1, 2>(), "arch"_a,
+        "log_level"_a = spdlog::level::to_short_c_str(defaultConfig.logLevel),
+        "max_filling_factor"_a = defaultConfig.schedulerConfig.maxFillingFactor,
+        "use_window"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.useWindow,
+        "window_min_width"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.windowMinWidth,
+        "window_ratio"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.windowRatio,
+        "window_share"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.windowShare,
+        "deepening_factor"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.deepeningFactor,
+        "deepening_value"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.deepeningValue,
+        "lookahead_factor"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.lookaheadFactor,
+        "reuse_level"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.reuseLevel,
+        "max_nodes"_a =
+            defaultConfig.layoutSynthesizerConfig.placerConfig.maxNodes,
+        "parking_offset"_a = defaultConfig.codeGeneratorConfig.parkingOffset,
+        "warn_unsupported_gates"_a =
+            defaultConfig.codeGeneratorConfig.warnUnsupportedGates);
+  }
   routingAwareCompiler.def_static(
       "from_json_string",
       [](const na::zoned::Architecture& arch,